+++ docker/tensorflow-aarch64/build.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 # *******************************************************************************
-# Copyright 2020-2022 Arm Limited and affiliates.
+# Copyright 2020-2023 Arm Limited and affiliates.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -39,7 +39,6 @@ function print_usage_and_exit {
   echo "                                 * tools          - build image including Python3 venv, with numpy."
   echo "                                 * dev            - build image including Bazel and TensorFlow, with sources."
   echo "                                 * tensorflow     - build image including TensorFlow build and benchmarks installed (default)."
-  echo "                                 * serving        - build image including TensorFlow Serving build."
   echo "                                 * full           - build all images."
   echo "      --build-target           AArch64 CPU target:"
   echo "                                 * generic        - portable build suitable for any ARMv8-A target (default)."
@@ -74,7 +73,6 @@ build_base_image=
 build_libs_image=
 build_tools_image=
 build_dev_image=
-build_serving_image=
 build_tensorflow_image=1
 readonly target_arch="aarch64"
 readonly host_arch=$(arch)
@@ -105,7 +103,6 @@ do
           build_libs_image=
           build_tools_image=
           build_dev_image=
-          build_serving_image=
           build_tensorflow_image=
           ;;
         libs )
@@ -113,7 +110,6 @@ do
           build_libs_image=1
           build_tools_image=
           build_dev_image=
-          build_serving_image=
           build_tensorflow_image=
           ;;
         tools )
@@ -121,7 +117,6 @@ do
           build_libs_image=
           build_tools_image=1
           build_dev_image=
-          build_serving_image=
           build_tensorflow_image=
           ;;
         dev )
@@ -129,15 +124,6 @@ do
           build_libs_image=
           build_tools_image=
           build_dev_image=1
-          build_serving_image=
-          build_tensorflow_image=
-          ;;
-        serving )
-          build_base_image=
-          build_libs_image=
-          build_tools_image=
-          build_dev_image=
-          build_serving_image=1
           build_tensorflow_image=
           ;;
         full )
@@ -145,7 +131,6 @@ do
           build_libs_image=1
           build_tools_image=1
           build_dev_image=1
-          build_serving_image=1
           build_tensorflow_image=1
           ;;
         tensorflow )
@@ -153,7 +138,6 @@ do
           build_libs_image=
           build_tools_image=
           build_dev_image=
-          build_serving_image=
           build_tensorflow_image=1
           ;;
         * )
@@ -245,15 +229,12 @@ if [[ $clean_build ]]; then
   extra_args="--pull --no-cache $extra_args"
 fi
 
-# Set TensorFlow and TensorFlow serving versions
-tf_version="v2.11.0"
-# Note: TF serving release tags do not have a `v` prefix
-tfserving_version="2.7.0"
+# Set TensorFlow version
+tf_version="v2.12.0"
 
 # Add build-args to pass version numbers,
 extra_args="$extra_args \
-    --build-arg tf_version=$tf_version \
-    --build-arg tfserving_version=$tfserving_version"
+    --build-arg tf_version=$tf_version"
 
 # Set CPU target props
 set_target $target
@@ -284,11 +265,6 @@ fi
 if [[ $build_dev_image ]]; then
   # Stage 4: Adds bazel and TensorFlow builds with sources and creates a whl.
   docker build $extra_args --target tensorflow-dev -t tensorflow-dev-v2$onednn:$image_tag .
-fi
-
-if [[ $build_serving_image ]]; then
-  # Stage 4b: Adds bazel and TensorFlow Serving builds with sources and creates a whl.
-  docker build $extra_args --target tensorflow-serving -t tensorflow-serving-v2$onednn:$image_tag .
 fi
 
 if [[ $build_tensorflow_image ]]; then