+++ docker/tensorflow-aarch64/CHANGELOG.md
@@ -10,10 +10,13 @@ where `YY` is the year, and `MM` the month of the increment.
 ### Added
 
 ### Changed
+- Updates TensorFlow to v2.12.
 
 ### Removed
+- Support for building TF Serving removed.
 
 ### Fixed
+- Make fixed format ACL convolution able to handle NCHW input.
 
 ## [r23.04] 2023-04-13
 https://github.com/ARM-software/Tool-Solutions/tree/tensorflow-pytorch-aarch64--r23.04/docker/tensorflow-aarch64